+++ application/includes/Mint/Mint.ArrayToXML.php
@@ -4,7 +4,7 @@
  *
  * @package		Mint
  * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
- * @copyright	Copyright (c) 2011, Squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
  * @license		GNU/GPL (General Public License)
  * @link		http://squallstar.it
  *