+++ application/includes/Mint/Mint.Rss.php
@@ -2,11 +2,11 @@
 /**
  * Mint Framework RSS
  *
- * @package		Mint
- * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
- * @copyright	Copyright (c) 2011, Squallstar
- * @license		GNU/GPL (General Public License)
- * @link		http://squallstar.it
+ * @package	Mint
+ * @author	Nicholas Valbusa - info@squallstar.it - @squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
+ * @license	GNU/GPL (General Public License)
+ * @link	http://squallstar.it
  *
  */