+++ application/javascript/application.js
@@ -3,7 +3,7 @@
  *
  * @package		Support-Ticketing
  * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
- * @copyright	Copyright (c) 2011, Squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
  * @license		GNU/GPL (General Public License)
  * @link		http://squallstar.it
  *
@@ -15,4 +15,8 @@ function sendMail(id, type, data) {
 
 function sendAsyncMail() {
 	$.get(root+'send-mail');
+}
+
+function switchAdvancedOptions() {
+	$('#advanced_options').slideToggle(300);
 }